adiciona uma solução de destroy
