+++ Platform.App/python-template/app/tests/controllers/test_command_controller.py
@@ -3,6 +3,7 @@ import pytest
 from mock import patch
 from utils.http import HttpClient, ExecutionResult
 import database
+from model.domain import conta
 import json
 
 
@@ -131,3 +132,26 @@ def test_full_persist(app):
         assert response.status_code == 200
         assert len(_list) == 1
         assert "id" in _list[0]
+
+def test_destroy_data(session, test_client):
+    origem = conta(titular="Fabio", saldo=10000)
+    destino = conta(titular="Moneda", saldo=100)
+    session.add_all([origem, destino])
+    session.commit()
+
+    with patch.object(HttpClient, 'get', return_value=apicore_map()) as mock_method:
+        uri = f'/Conta/Conta?filter=transferencia&origem={origem.id}&destino={destino.id}'
+        status_code, resp = test_client.get_json(uri)
+        destroyed_id = resp[0]["id"]
+        resp[0]["_metadata"]["changeTrack"] = "destroy"
+        body = json.dumps([resp[0]])
+        assert len(resp) == 2
+        response = test_client.post('/Conta/persist',follow_redirects=True, data=body,
+                               content_type='application/json')
+
+        assert response.status_code == 200
+
+        status_code, resp = test_client.get_json(uri)
+        assert status_code == 200
+        assert len(resp) == 1
+        assert resp[0]["id"] != destroyed_id
